+++ lsql-core/src/main/java/com/w11k/lsql/converter/predefined/AtomicIntegerConverter.java
@@ -16,12 +16,12 @@ public class AtomicIntegerConverter extends Converter {
     }
 
     @Override
-    protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
+    public void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
         ps.setInt(index, ((AtomicInteger) val).get());
     }
 
     @Override
-    protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
+    public Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         return new AtomicInteger(rs.getInt(index));
     }