+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/BlobConverter.java
@@ -12,13 +12,13 @@ public class BlobConverter extends Converter {
     }
 
     @Override
-    protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
+    public void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
         com.w11k.lsql.Blob blob = (com.w11k.lsql.Blob) val;
         ps.setBlob(index, blob.getInputStream());
     }
 
     @Override
-    protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
+    public Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         Blob blob = rs.getBlob(index);
         return new com.w11k.lsql.Blob(blob.getBinaryStream());
     }