+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/BooleanConverter.java
@@ -19,12 +19,12 @@ public class BooleanConverter extends Converter {
     }
 
     @Override
-    protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
+    public void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
         ps.setBoolean(index, (Boolean) val);
     }
 
     @Override
-    protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
+    public Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         return rs.getBoolean(index);
     }
 }