+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/JodaDateTimeConverter.java
@@ -13,14 +13,14 @@ public class JodaDateTimeConverter extends Converter {
     }
 
     @Override
-    protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
+    public void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
         DateTime dt = (DateTime) val;
         Timestamp ts = new Timestamp(dt.getMillis());
         ps.setTimestamp(index, ts);
     }
 
     @Override
-    protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
+    public Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         Timestamp timestamp = rs.getTimestamp(index);
         if (timestamp != null) {
             return new DateTime(timestamp.getTime());