+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/JodaLocalDateConverter.java
@@ -13,14 +13,14 @@ public class JodaLocalDateConverter extends Converter {
     }
 
     @Override
-    protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
+    public void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
         LocalDate dt = (LocalDate) val;
         Timestamp ts = new Timestamp(dt.toDate().getTime());
         ps.setTimestamp(index, ts);
     }
 
     @Override
-    protected Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
+    public Object getValue(LSql lSql, ResultSet rs, int index) throws SQLException {
         Timestamp timestamp = rs.getTimestamp(index);
         if (timestamp != null) {
             return new LocalDate(timestamp.getTime());